Updated Glossary and references in Use Case Files
Updated the glossary entries, migrating organizations + adding two columns (References & Type).
Updated use case descriptions to use the acronyms included in the Glossary.
Added a Release Note box to the top of Glossary to indicate that the table is a "sink" for this version but that the content may be sorted out to other tables and sections in later versions.