SDPi 1.0 dev internal review v2 (#303)
* Added the SDPi 1.0 DEV Internal Review v1 HTML files

HTML files for DEV Internal review (ver. 1)

* Issue #256 Typos Corrected

Corrected the typos identified in Issue #256

* SDPi 1.0 Comment Resolutions  - From Henderson NV 2023.01.19

Fixed Transaction table column/row alignment.
Typos identified in #269

* Adjusted TF-1 Offsets to Sync Depth

Normalizing the TF-1 section offset & depth to be sync'd with those of TF-2.  This should address the challenges around the missing TF-1 figure & section numbering + the ToC outline depth.

* Additional TF-1 Section Numbering Adjustments

Continuing to adjust the offset & depth for the TF-1 Table of Contents to reflect what is in TF-2 + the Word draft of the supplement.

* TF-0 & Supplement Overview

Adjusting section designations to add display of content to the Table of Contents at the top.

* Adjusting TF-0 and Supplement Intro Sections

Adjusting Table of Contents per TF-1 to -3 organization.

* Updated cover to sync with IUA

Updated the title page content and organization to sync with the IUA (https://profiles.ihe.net/ITI/IUA/index.html#) profile.

* Comment Resolution Updates - 2023.01.26

Including those comments submitted by Javier ...
> Indented TF-0 Appendices under the "General Introduction Appendices" section
> MDIB Report Retrofit / Best Practices #270
> SDPi-xC with Mixed Device Safety Classes #271
> #272, #273, #275, #277, #279, #280, #281, #282, #283

* Tweak "cover page".

* Removed generated HTML files from sdpi-supplement folder

Removed the files that are generated for HTML publication from the sdpi-supplement folder.

---------

Co-authored-by: David Gregorczyk <david.gregorczyk@web.de>